+++ uui-timeline/src/draw/scale.ts
@@ -38,6 +38,7 @@ const defaultColors = {
 
 const defaultWidth = {
     cellBorderWidth: 1,
+    todayLineHeight: 4,
 };
 
 const moveAmount = 0.7;
@@ -379,10 +380,16 @@ const drawHours = ({
         });
 };
 
-const drawToday = ({ context, scaleBar, todayLineColor = defaultColors.todayLineColor }: CanvasDrawHeaderTodayProps) => {
+const drawToday = ({
+    context,
+    scaleBar,
+    todayLineColor = defaultColors.todayLineColor,
+    canvasHeight,
+    todayLineHeight = defaultWidth.todayLineHeight,
+}: CanvasDrawHeaderTodayProps) => {
     if (isCurrentPeriod(scaleBar.leftDate, scaleBar.rightDate)) {
         context.fillStyle = todayLineColor;
-        context.fillRect(scaleBar.left + 1, 56, scaleBar.right - scaleBar.left, 4);
+        context.fillRect(scaleBar.left + 1, canvasHeight - todayLineHeight, scaleBar.right - scaleBar.left, todayLineHeight);
     }
 };
 
@@ -393,6 +400,7 @@ const drawTopDays = ({
     topDayTextColor = defaultColors.topDayTextColor,
     weekendTextColor = defaultColors.weekendTextColor,
     todayLineColor = defaultColors.todayLineColor,
+    todayLineHeight = defaultWidth.todayLineHeight,
     drawToday: customDrawToday,
     getTopMonth: customGetMonth = getTopMonth,
     canvasHeight,
@@ -421,7 +429,7 @@ const drawTopDays = ({
             textColor,
             ...restProps,
         });
-        (customDrawToday ?? drawToday)({ context, scaleBar: w, todayLineColor });
+        (customDrawToday ?? drawToday)({ context, scaleBar: w, todayLineColor, todayLineHeight, canvasHeight });
         drawBorderForTopCell({ context, canvasHeight, scaleBar: w, width: cellBorderWidth, color: cellBorderColor });
     });
 };
@@ -433,6 +441,7 @@ const drawDays = ({
     periodTextColor = defaultColors.periodTextColor,
     weekendTextColor = defaultColors.weekendTextColor,
     todayLineColor = defaultColors.todayLineColor,
+    todayLineHeight = defaultWidth.todayLineHeight,
     drawToday: customDrawToday,
     canvasHeight,
     cellBorderColor = defaultColors.cellBorderColor,
@@ -447,7 +456,7 @@ const drawDays = ({
         const color = isHoliday ? weekendCellBackgroundColor : cellBackgroundColor;
         drawCellBackground({ context, scaleBar: w, canvasHeight, y: getCanvasVerticalCenter(canvasHeight), color });
 
-        (customDrawToday ?? drawToday)({ context, scaleBar: w, todayLineColor });
+        (customDrawToday ?? drawToday)({ context, scaleBar: w, todayLineColor, todayLineHeight, canvasHeight });
         drawBottomGridLine({ context, canvasHeight, scaleBar: w, width: cellBorderWidth, color: cellBorderColor });
 
         const text = w.leftDate.getDate().toString();
@@ -510,6 +519,7 @@ const drawWeeks = ({
     visibility,
     periodTextColor = defaultColors.periodTextColor,
     todayLineColor = defaultColors.todayLineColor,
+    todayLineHeight = defaultWidth.todayLineHeight,
     drawToday: customDrawToday,
     canvasHeight,
     cellBorderColor = defaultColors.cellBorderColor,
@@ -523,7 +533,7 @@ const drawWeeks = ({
         const isCurPeriod = isCurrentPeriod(w.leftDate, w.rightDate);
         drawCellBackground({ context, scaleBar: w, canvasHeight, y: getCanvasVerticalCenter(canvasHeight), color: cellBackgroundColor });
 
-        (customDrawToday ?? drawToday)({ context, scaleBar: w, todayLineColor });
+        (customDrawToday ?? drawToday)({ context, scaleBar: w, todayLineColor, todayLineHeight, canvasHeight });
         drawBottomGridLine({ context, canvasHeight, scaleBar: w, width: cellBorderWidth, color: cellBorderColor });
         drawPeriodText({
             context,
@@ -546,6 +556,7 @@ const drawBottomMonths = ({
     canvasHeight,
     periodTextColor = defaultColors.periodTextColor,
     todayLineColor = defaultColors.todayLineColor,
+    todayLineHeight = defaultWidth.todayLineHeight,
     drawToday: customDrawToday,
     cellBorderColor = defaultColors.cellBorderColor,
     cellBorderWidth = defaultWidth.cellBorderWidth,
@@ -569,7 +580,7 @@ const drawBottomMonths = ({
             textColor: periodTextColor,
             ...restProps,
         });
-        (customDrawToday ?? drawToday)({ context, scaleBar: w, todayLineColor });
+        (customDrawToday ?? drawToday)({ context, scaleBar: w, todayLineColor, todayLineHeight, canvasHeight });
         drawBottomGridLine({ context, canvasHeight, scaleBar: w, width: cellBorderWidth, color: cellBorderColor });
     });
 };
@@ -580,6 +591,7 @@ const drawYears = ({
     visibility,
     periodTextColor = defaultColors.periodTextColor,
     todayLineColor = defaultColors.todayLineColor,
+    todayLineHeight = defaultWidth.todayLineHeight,
     drawToday: customDrawToday,
     canvasHeight,
     cellBorderColor = defaultColors.cellBorderColor,
@@ -599,7 +611,7 @@ const drawYears = ({
             const y = canvasHeight;
             timelinePrimitives.drawHorizontalLine({ context, x1: w.left, x2: w.right + 1, y: y - 1, color: cellBorderColor, width: cellBorderWidth });
             drawCellBackground({ context, scaleBar: w, canvasHeight, height: canvasHeight, color: cellBackgroundColor });
-            (customDrawToday ?? drawToday)({ context, scaleBar: w, todayLineColor });
+            (customDrawToday ?? drawToday)({ context, scaleBar: w, todayLineColor, todayLineHeight, canvasHeight });
             timelinePrimitives.drawVerticalLine({ context, x: w.left + 0.5, y2: y - 1, color: cellBorderColor, width: cellBorderWidth });
         } else {
             const color = w.leftDate.getFullYear() % 2 === 0