+++ changelog.md
@@ -8,6 +8,7 @@
 * [Tag]: changed layout - added gaps between internal items, changed padding
 * [Data Sources]: cursor-based pagination support. More details [here](http://uui.epam.com/documents?id=dataSources-lazy-dataSource&mode=doc&category=dataSources&theme=loveship#using_cursor-based_pagination)
 * [TimelineScale]: added bottom/top month text customisation.
+* [TimelineScale]: customisation of today line height was added.
 
 **What's Fixed**
 * [VirtualList]: fixed estimatedHeight calculations in components with pagination